Optical-Character-Recognition-App/requirements.txt
opencv-python-headless==4.2.0.32
pytesseract==0.3.7
numpy==1.19.3
Flask==1.1.1
Pillow==8.0.1
